Curso 06 · Aula 08 · Atividade 04 · Finalmente Lambdas

+++ curso-06-java-parte-06-conhecendo-o-java-util/src/br/com/bytebank/system/model/Account.java
@@ -147,7 +147,7 @@ public abstract class Account implements Comparable<Account> {
 	
 	@Override
 	public int compareTo(Account account) {
-		return Double.compare(this.accountNumber, account.accountNumber);
+		return Integer.compare(this.accountNumber, account.accountNumber);
 	}
 	
 	@Override
